requirements.txt
inquirer==3.1.3
loralib==0.1.1
ninja==1.11.1
pynvml==11.5.0
scipy==1.10.1
sentencepiece==0.1.97
simple-ai-server==0.2.1
tokenizers==0.13.2
torch==2.0.0
transformers==4.27.3
git+https://github.com/Nintorac/ChatRWKV.git@parallel#subdirectory=rwkv_pip_package 
aiorwlock
ray[default]
ray[serve]
gpustat==1.0.0
kubernetes